Extract remaining runtime host hotspots
