+++ desktop/runtime-host/live-runtime-registry.ts
@@ -1,30 +1,18 @@
 import path from 'node:path'
-import { normalizeModelRegistryContextWindows } from '../../shared/model-context-window-normalization.ts'
 import { getPersistedSessionPath } from '../../shared/session-paths.ts'
-import { ensureAskQuestionsExtensionRuntimePath } from '../native-extensions/ask-questions-extension-path.ts'
 import { getPiModule } from '../pi-module.ts'
-import {
-  abortHeadlessExtensionCommand,
-  isHeadlessExtensionCommandRunning,
-} from '../runtime/agent-session-extensions.ts'
-import { createArtifactTools } from '../runtime/artifact-tools.ts'
-import { createAttachmentFileTools } from '../runtime/attachment-file-tools.ts'
-import { getRuntimeSystemPrompt } from '../runtime/chat-system-prompt.ts'
 import { buildComposerState } from '../runtime/composer-state.ts'
-import {
-  createIsolatedRuntimeResourceLoader,
-  createRuntimeSettingsManager,
-} from '../runtime/isolated-settings-manager.ts'
 import { normalizeRuntimeSettingsCwd } from '../runtime/runtime-settings-cwd.ts'
 import type { PiRuntime } from '../runtime/types.ts'
-import { publishComposerUpdate } from './live-thread-publisher.ts'
-import { invokeMainRequest } from './main-request-client.ts'
-import { createNativeAskQuestionsTools } from './native-ask-questions-tool.ts'
 import {
-  bindRuntimeExtensionHandlers,
-  refreshRuntimeExtensionHandlers,
-} from './runtime-extension-bindings.ts'
-import { handleRuntimeSessionEvent } from './runtime-session-events.ts'
+  abortRuntimeExtensionCommand,
+  createLiveRuntime,
+  isRuntimeExtensionCommandRunning,
+  refreshRuntimeExtensionBindings as refreshRuntimeExtensionBindingsWithReload,
+} from './live-runtime-factory.ts'
+import { publishComposerUpdate } from './live-thread-publisher.ts'
+
+export { abortRuntimeExtensionCommand, isRuntimeExtensionCommandRunning }
 
 type RuntimeRecord = {
   runtimePromise: Promise<PiRuntime>
@@ -37,6 +25,16 @@ const RUNTIME_IDLE_TIMEOUT_MS = 15 * 60 * 1_000
 const runtimeRecords = new Map<string, RuntimeRecord>()
 const runtimeMutationTails = new Map<string, Promise<void>>()
 const staleRuntimeGenerations = new Map<string, number>()
+
+const liveRuntimeFactoryHandlers = {
+  reloadRuntimeSettingsIfSafe,
+  scheduleRuntimeDisposal,
+  suspendRuntimeDisposal,
+}
+
+export async function refreshRuntimeExtensionBindings(runtime: PiRuntime) {
+  await refreshRuntimeExtensionBindingsWithReload(runtime, reloadRuntimeSettingsIfSafe)
+}
 
 async function disposeRuntimeIfIdle(runtimeKey: string, record: RuntimeRecord) {
   const currentRecord = runtimeRecords.get(runtimeKey)
@@ -76,164 +74,6 @@ export function scheduleRuntimeDisposal(runtimeKey: string) {
   record.disposeTimeout = setTimeout(() => {
     void disposeRuntimeIfIdle(runtimeKey, record)
   }, RUNTIME_IDLE_TIMEOUT_MS)
-}
-
-async function getEnabledNativeExtensionsForRuntime(options: {
-  sessionManager?: PiRuntime['session']['sessionManager']
-}) {
-  const sessionPath = options.sessionManager?.getSessionFile?.() ?? null
-  if (sessionPath) {
-    const enabled = await invokeMainRequest('getSessionNativeExtensions', { sessionPath })
-    if (enabled) return enabled
-    const defaultEnabled = await invokeMainRequest('snapshotDefaultNativeExtensions', {})
-    await invokeMainRequest('setSessionNativeExtensions', {
-      sessionPath,
-      enabled: defaultEnabled,
-    })
-    return defaultEnabled
-  }
-
-  return await invokeMainRequest('snapshotDefaultNativeExtensions', {})
-}
-
-async function createRuntime(options: {
-  cwd: string
-  sessionDir?: string | undefined | null | undefined
-  settingsCwd?: string | undefined | null | undefined
-  chatGroupId?: string | undefined | null | undefined
-  sessionManager?: PiRuntime['session']['sessionManager']
-}): Promise<PiRuntime> {
-  const {
-    AuthStorage,
-    ModelRegistry,
-    SessionManager,
-    SettingsManager,
-    DefaultResourceLoader,
-    createAgentSession,
-    defineTool,
-    getAgentDir,
-  } = await getPiModule()
-  const agentDir = getAgentDir()
-  const authStorage = AuthStorage.create()
-  const modelRegistry = normalizeModelRegistryContextWindows(
-    ModelRegistry.create(authStorage, `${agentDir}/models.json`),
-  )
-  const settingsManager = createRuntimeSettingsManager({
-    SettingsManager,
-    cwd: options.cwd,
-    agentDir,
-    settingsCwd: options.settingsCwd,
-  })
-  const sessionDir = options.sessionDir ?? settingsManager.getSessionDir() ?? undefined
-  const resourceLoader = await createIsolatedRuntimeResourceLoader({
-    DefaultResourceLoader,
-    cwd: options.cwd,
-    agentDir,
-    settingsCwd: options.settingsCwd,
-    settingsManager,
-    systemPrompt: getRuntimeSystemPrompt({ settingsCwd: options.settingsCwd }),
-  })
-  let runtime: PiRuntime | null = null
-  const enabledNativeExtensions = await getEnabledNativeExtensionsForRuntime(
-    options.sessionManager ? { sessionManager: options.sessionManager } : {},
-  )
-  const nativeAskQuestionTools = enabledNativeExtensions.includes('askQuestions')
-    ? await createNativeAskQuestionsTools({
-        defineTool,
-        extensionPath: ensureAskQuestionsExtensionRuntimePath() ?? '',
-        getRuntime: () => runtime,
-        onStateChange: () => {
-          if (!runtime) return
-          const activeRuntime = runtime
-          void buildComposerState(activeRuntime).then((composer) => {
-            publishComposerUpdate(composer, {
-              projectId: activeRuntime.cwd,
-              sessionPath: activeRuntime.session.sessionFile,
-            })
-          })
-        },
-      })
-    : []
-  const attachmentFileTools = options.settingsCwd
-    ? createAttachmentFileTools({
-        cwd: options.cwd,
-        autoResizeImages: settingsManager.getImageAutoResize(),
-      })
-    : null
-  const { session } = await createAgentSession({
-    cwd: options.cwd,
-    agentDir,
-    authStorage,
-    modelRegistry,
-    settingsManager,
-    ...(resourceLoader ? { resourceLoader } : {}),
-    sessionManager: options.sessionManager ?? SessionManager.create(options.cwd, sessionDir),
-    ...(options.settingsCwd
-      ? {
-          noTools: 'builtin' as const,
-          customTools: [
-            ...(attachmentFileTools?.tools ?? []),
-            ...createArtifactTools({
-              createArtifact: (input) => invokeMainRequest('createArtifact', input),
-              editArtifact: (input) => invokeMainRequest('editArtifact', input),
-              getArtifact: ({ conversationId, slug }) =>
-                invokeMainRequest('getArtifact', { artifactSlug: slug, conversationId }),
-              listArtifacts: (conversationId) =>
-                invokeMainRequest('listArtifacts', { conversationId }),
-            }),
-            ...nativeAskQuestionTools,
-          ],
-        }
-      : nativeAskQuestionTools.length > 0
-        ? { customTools: nativeAskQuestionTools }
-        : {}),
-  })
-  runtime = {
-    cwd: options.cwd,
-    session,
-    chatGroupId: options.chatGroupId ?? null,
-    attachmentFileAccess: attachmentFileTools?.access,
-  } satisfies PiRuntime
-
-  if (!options.sessionManager) {
-    const runtimeKey = getPersistedSessionPath(runtime.session.sessionFile)
-    if (runtimeKey) {
-      await invokeMainRequest('setSessionNativeExtensions', {
-        sessionPath: runtimeKey,
-        enabled: enabledNativeExtensions,
-      })
-    }
-  }
-
-  session.subscribe((event) =>
-    handleRuntimeSessionEvent(runtime, event, {
-      isRuntimeExtensionCommandRunning,
-      reloadRuntimeSettingsIfSafe,
-      scheduleRuntimeDisposal,
-      suspendRuntimeDisposal,
-    }),
-  )
-
-  await bindRuntimeExtensionHandlers(runtime, {
-    isRuntimeExtensionCommandRunning,
-    reloadRuntimeSettingsIfSafe,
-  })
-  return runtime
-}
-
-export function abortRuntimeExtensionCommand(runtime: PiRuntime) {
-  return abortHeadlessExtensionCommand(runtime.session)
-}
-
-export function isRuntimeExtensionCommandRunning(runtime: PiRuntime) {
-  return isHeadlessExtensionCommandRunning(runtime.session)
-}
-
-export async function refreshRuntimeExtensionBindings(runtime: PiRuntime) {
-  await refreshRuntimeExtensionHandlers(runtime, {
-    isRuntimeExtensionCommandRunning,
-    reloadRuntimeSettingsIfSafe,
-  })
 }
 
 function registerRuntime(
@@ -284,12 +124,15 @@ export async function getOrCreateRuntimeForSessionPath(
   const { SessionManager } = await getPiModule()
   const sessionManager = SessionManager.open(persistedSessionPath)
   let record: RuntimeRecord | null = null
-  const runtimePromise = createRuntime({
-    cwd: sessionManager.getCwd(),
-    settingsCwd,
-    chatGroupId: options.chatGroupId ?? null,
-    sessionManager,
-  }).catch((error) => {
+  const runtimePromise = createLiveRuntime(
+    {
+      cwd: sessionManager.getCwd(),
+      settingsCwd,
+      chatGroupId: options.chatGroupId ?? null,
+      sessionManager,
+    },
+    liveRuntimeFactoryHandlers,
+  ).catch((error) => {
     if (record && runtimeRecords.get(persistedSessionPath) === record)
       runtimeRecords.delete(persistedSessionPath)
     staleRuntimeGenerations.delete(persistedSessionPath)
@@ -304,12 +147,15 @@ export async function createRuntimeForNewSession(
   sessionDir?: string | undefined | null | undefined,
   options: { chatGroupId?: string | undefined | null | undefined } = {},
 ) {
-  const runtime = await createRuntime({
-    cwd,
-    sessionDir,
-    settingsCwd: sessionDir ?? null,
-    chatGroupId: options.chatGroupId ?? null,
-  })
+  const runtime = await createLiveRuntime(
+    {
+      cwd,
+      sessionDir,
+      settingsCwd: sessionDir ?? null,
+      chatGroupId: options.chatGroupId ?? null,
+    },
+    liveRuntimeFactoryHandlers,
+  )
   const runtimeKey = getPersistedSessionPath(runtime.session.sessionFile)
   if (runtimeKey) {
     registerRuntime(runtimeKey, Promise.resolve(runtime), sessionDir ?? null)
